nuke exynos-argos
Argos is a driver that adjusts the performance of the SoC according to the throughput. However, it is ineffective and only causes unnecessary throttling of the device. According to several intensive tests, the SoC performed about 4% better without Argos.

There is still work to be done, such as removing some or all dependency on components in pm_qos, so modifications have been made to the QoS area. For example, some frequencies are now forced directly without using the QoS request.

Signed-off-by: Elchanz3 <inutilidades639@gmail.com>